requirements.txt
discord.py>=2.3.2
yt-dlp==2023.10.13
spotipy==2.23.0
google-api-python-client==2.104.0
Flask==2.3.2
python-dotenv==1.0.0